Hybrid_RAG/requirements.txt
langchain>=0.2.0
langchain-chroma
langchain-google-genai
langchain-huggingface
langchain-community
sentence-transformers
langchain-neo4j
python-dotenv
wikipedia
neo4j
pydantic>=1.10.0
langchain_experimental
ragas
tabulate
